Redesign home page and change background
