+++ pages/404.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "firebase/firestore";
 
-export default function CertPage() {
+export default function NotFound() {
     return (
         <>Page Not Found</>
     );
