+++ styles/Home.module.css
@@ -22,12 +22,9 @@
 .footer {
   width: 100%;
   height: 80px;
-  border-top: 1px solid #eaeaea;
   display: flex;
   justify-content: center;
   align-items: center;
-
-  background-color: #353535;
 }
 
 .title a {
@@ -55,7 +52,7 @@
 
 .description {
   line-height: 1.5;
-  font-size: 1.5rem;
+  font-size: 2rem;
 }
 
 .grid {
@@ -73,19 +70,24 @@
   text-align: left;
   color: inherit;
   text-decoration: none;
-  border: 1px solid #eaeaea;
-  border-radius: 10px;
   transition: color 0.15s ease, border-color 0.15s ease;
   width: 45%;
 
-  background-color: #353535;
+  background-color: #f2f2f2;
+  border-radius: 18px;
+  border: 1px solid black;
+	outline: black solid 1px;
+
+  color: black;
+
 }
 
 .card:hover,
 .card:focus,
 .card:active {
-  color: #3c93f7;
+  color: #0e2aa8;
   border-color: #3c93f7;
+  outline: none;
 }
 
 .card h2 {
