+++ styles/globals.css
@@ -4,6 +4,9 @@ body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
     Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+
+  background: url('../public/bg-fabric.png');
+  /* background-size: cover; */
 }
 
 .container {
@@ -14,7 +17,6 @@ body {
   justify-content: center;
   align-items: center;
 
-  background-color: #474747;
   color: white;
 }
